refactor: Remove WispPresented protocol and move its methods into UIViewController extension
